benchmarks/algorithms/run_iscale/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
Bottleneck==1.4.2
cffi==2.0.0
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
hdbscan==0.8.40
idna==3.11
imagecodecs==2025.3.30
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
lightning-utilities==0.15.2
llvmlite==0.45.1
MarkupSafe==3.0.3
matplotlib==3.10.7
mkl-service==2.4.0
mkl_fft==1.3.11
mkl_random==1.2.8
mpmath==1.3.0
multidict==6.7.0
networkx==3.4.2
numba==0.62.1
numexpr==2.11.0
numpy==2.0.1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
propcache==0.4.1
pycparser==2.23
pynndescent==0.5.13
pyparsing==3.2.5
python-dateutil==2.9.0post0
pytorch-lightning==2.5.5
pytz==2025.2
pyvips==3.0.0
PyYAML==6.0.3
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.8.0
torchmetrics==1.8.2
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
yarl==1.22.0
